backend/requirements.txt
fastapi
uvicorn[standard]
aiortc
torch
torchvision
Pillow
whisper
langchain
langgraph
openai
pydantic
